Move extensions to proper folder | move request from view to flutter view
