+++ ios/Runner/Extensions/Foundation+Extensions/Collection+Extensions/Collection+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  Collection+Extensions.swift
+//  Runner
+//
+
+import Foundation
+
+extension Collection {
+		/// Returns the element at the specified index if it is within bounds, otherwise nil.
+		subscript (safe index: Index) -> Element? {
+				return indices.contains(index) ? self[index] : nil
+		}
+}
